Restrict cross-compartment tail-call optimization.
If the following is the control flow: C1 -calls> C254 -calls> C1 -tailcalls> C4. Now from C4 when it returns directly to C254's stub, then the context of C254 is not restored properly.
When C1 goes to C4, CSR_UCURRCAP is changed, but is not restored back when it returns to C254. On return to C254, CSR_UCURRCAP remains as 4 and is not changed back to 1. Either we can support this case
by storing the CSR_CURRCAP on stack even on 254 compartment switch. But this might affect the overhead of all 254 switches. So for now, disabling the cross-compartment tail-call optimization for this scenario.