Change C232 to C230
Reason being camlIpaddr__entry belonging to C232 is placed within C231 due to some expression camlIpaddr__* present in C231 compartment. Because of this expression, camlIpaddr__entry is also placed in C231 and the base and bound of of C232 is not configured properly due to this. Solution: 1). Find a way to escape special characters in linker files. 2). Use some other compartment before C231 to place camlIpaddr_entry() so that it is not placed within C231.